+++ lib/fetching/bot-master.js
@@ -239,7 +239,7 @@ BotMaster.prototype._add = function(bot, callback) {
     bot.contentService.reloadSettings();
   });
   this.on('ctListUpdated', function() {
-    if(bot.source.media === 'crowdtangle'){
+    if(bot.source.media === 'facebook' || bot.source.media === 'instagram'){
       bot.contentService.reloadCTList();
     }
   });
