+++ lib/fetching/content-service-factory.js
@@ -14,7 +14,8 @@ contentServices['dummy-fast'] = require('./content-services/dummy-fast-content-s
 contentServices.smsgh = require('./content-services/smsgh-content-service');
 contentServices.whatsapp = require('./content-services/whatsapp-content-service');
 contentServices.replay = require('./content-services/replay-content-service');
-contentServices.crowdtangle = require('./content-services/crowd-tangle-content-service');
+contentServices.facebook = require('./content-services/facebook-content-service');
+contentServices.instagram = require('./content-services/instagram-content-service');
 contentServices.comments = require('./content-services/comments-content-service');
 
 function ContentServiceFactory() { /* empty constructor */ }
