+++ models/report.js
@@ -102,7 +102,7 @@ schema.methods.addSMTCTag = function(smtcTagId, callback) {
     this.smtcTags.push({_id: smtcTagId});
     this.read = true;
     // Only send a post to the acquisition API if it is a) not a comment b) a FB post and c) not a group post
-    if (!this.commentTo && this._media[0] === 'crowdtangle' && !this.url.match(/permalink/)) {
+    if (!this.commentTo && this._media[0] === 'facebook' && !this.url.match(/permalink/)) {
       SMTCTag.findById(smtcTagId, (err, tag) => {
         if (err) {
           logger.error(err);
@@ -133,7 +133,7 @@ schema.methods.removeSMTCTag = function(smtcTagId, callback) {
     if (fndIndex !== -1) {
       this.smtcTags.splice(fndIndex, 1);
 
-      if (!this.commentTo && this._media[0] === 'crowdtangle') {
+      if (!this.commentTo && this._media[0] === 'facebook') {
         SMTCTag.findById(smtcTagId, (err, tag) => {
           if (err) {
             logger.error(err);
