+++ public/angular/js/app.js
@@ -145,7 +145,6 @@ require('./routes');
 // Filters
 require('./filters/aggie-date');
 require('./filters/capitalize');
-require('./filters/replaceCT');
 require('./filters/delay');
 require('./filters/interval');
 require('./filters/max-count');
