+++ public/angular/js/controllers/incidents/report_incident_modal.js
@@ -155,8 +155,9 @@ angular.module('Aggie')
       // deleted
 
       if (report.metadata.platform === "Facebook") {
-        // set Facebook as source for CrowdTangle reports
         return 'facebook-source';
+      } else if (report.metadata.platform === "Instagram") {
+        return 'instagram-source';
       } else {
         for (var i = 0; i < report._sources.length; i++) {
           var sourceId = report._sources[i];
