+++ public/angular/js/controllers/sources/index.js
@@ -31,12 +31,12 @@ angular.module('Aggie')
     };
 
     $scope.target = function(source) {
-      return source.media == 'twitter' || source.media == 'smsgh' || source.media == 'whatsapp' || source.media == 'crowdtangle' || source.media == 'comments' ? source.keywords : source.url;
+      return source.media == 'twitter' || source.media == 'smsgh' || source.media == 'whatsapp' || source.media == 'facebook' || source.media == 'instagram' || source.media == 'comments' ? source.keywords : source.url;
     };
 
     $scope.sourceClass = function(source) {
 
-      var mediaOptions = ['twitter', 'rss', 'elmo', 'smsgh', 'whatsapp', 'crowdtangle', 'comments'];
+      var mediaOptions = ['twitter', 'rss', 'elmo', 'smsgh', 'whatsapp', 'facebook', 'instagram', 'comments'];
 
       if (mediaOptions.indexOf(source.media) !== -1) {
         return source.media + '-source';
