+++ public/angular/sass/components/tables.scss
@@ -119,6 +119,9 @@ tbody {
   .facebook-source {
       background: url('/images/facebook.png') no-repeat 0 50%;
   }
+  .instagram-source {
+    background: url('/images/instagram.png') no-repeat 0 50%;
+  }
   .twitter-source {
     background: transparent url('/images/twitter.png') no-repeat 0 50%;
     &:before {
