+++ public/angular/templates/api-settings-modal.html
@@ -2,7 +2,7 @@
   <form name="settingsForm" class="form-horizontal" novalidate>
     <div class="modal-header">
       <h5 class="modal-title">
-        <div class="icon-left" translate="api_settings_modal.title" translate-value-api="{{ item | replaceCT }}"></div>
+        <div class="icon-left" translate="api_settings_modal.title" translate-value-api="{{ item }}"></div>
       </h5>
       <button class="close" ng-click="close()" aria-hidden="true">×</button>
     </div>
