+++ public/angular/templates/config.html
@@ -44,7 +44,7 @@
                 <tbody>
                   <tr ng-repeat="settingOption in apiSettingOptions">
                     <td>
-                      <label class="settings-label">{{ settingOption| replaceCT | translate }}</label>
+                      <label class="settings-label">{{ settingOption | translate }}</label>
                     </td>
                     <td class="action text-center">
                       <div class="ng-scope">
