+++ public/angular/templates/reports/content-services/facebook-show.html
@@ -1,0 +1,98 @@
+<div ng-include="'/templates/reports/content-services/crowdtangle.html'"></div>
+<div class="row">
+    <div class="col">
+    <h5>Statistics</h5>
+    <div class="table-responsive">
+        <table class="table table-sm">
+        <tr>
+            <th translate>Real Statistics</th>
+            <td class="detail-value">
+        <span title="{{'Comments' | translate}}">
+            <strong>{{ "Comments:" | translate }}</strong> {{report.metadata.actualStatistics.commentCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Shares' | translate}}">
+            <strong>{{ "Shares:" | translate }}</strong> {{report.metadata.actualStatistics.shareCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Likes' | translate}}">
+            <b>{{'Likes:' | translate}}</b> {{report.metadata.actualStatistics.likeCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Love Reacts' | translate}}">
+            <b>❤'s</b> {{report.metadata.actualStatistics.loveCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Haha Reacts' | translate}}">
+            <b>😂's:</b> {{report.metadata.actualStatistics.hahaCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Wow Reacts' | translate}}">
+            <b>😯's</b> {{report.metadata.actualStatistics.wowCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Sad Reacts' | translate}}">
+            <b>😢's: </b>{{report.metadata.actualStatistics.sadCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Angry Reacts' | translate}}">
+            <b>😡's: </b>{{report.metadata.actualStatistics.angryCount}}
+        </span>
+            </td>
+        </tr>
+        <!---- Crowdtangle Expected Statistics ---->
+        <tr>
+            <th translate>Expected Statistics</th>
+            <td class="detail-value">
+        <span title="{{'Comments' | translate}}">
+            <strong>{{ "Comments:" | translate }}</strong> {{report.metadata.expectedStatistics.commentCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Shares' | translate}}">
+            <strong>{{ "Shares:" | translate }}</strong> {{report.metadata.expectedStatistics.shareCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Likes' | translate}}">
+            <b>{{'Likes:' | translate}}</b> {{report.metadata.expectedStatistics.likeCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Love Reacts' | translate}}">
+            <b>❤'s</b> {{report.metadata.expectedStatistics.loveCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Haha Reacts' | translate}}">
+            <b>😂's:</b> {{report.metadata.expectedStatistics.hahaCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Wow Reacts' | translate}}">
+            <b>😯's</b> {{report.metadata.expectedStatistics.wowCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Sad Reacts' | translate}}">
+            <b>😢's: </b>{{report.metadata.expectedStatistics.sadCount}}
+        </span>
+            </td>
+            <td class="detail-value">
+        <span title="{{'Angry Reacts' | translate}}">
+            <b>😡's: </b>{{report.metadata.expectedStatistics.angryCount}}
+        </span>
+            </td>
+        </tr>
+        <!------------------ End Crowdtangle Data ------------------>
+        </table>
+    </div>
+    </div>
+</div>
