+++ public/angular/templates/sources/modal.html
@@ -18,7 +18,7 @@
         <!-- Initial Source Media Choice -->
         <div class="form-group" ng-class="{ 'read-only': source._id }">
           <label for="media" translate>Media</label>
-          <select id="media" class="form-control" ng-model="source.media" ng-options="media as media| replaceCT | translate for media in mediaOptions" required ng-hide="source._id">
+          <select id="media" class="form-control" ng-model="source.media" ng-options="media as media | translate for media in mediaOptions" required ng-hide="source._id">
             <option value="" translate>Choose source media</option>
           </select>
           <input type="text" class="form-control" ng-model="source.media" ng-show="source._id" readonly>
@@ -43,7 +43,7 @@
           </div>
         </div>
 
-        <div class="form-group" ng-show="source.media == 'smsgh' || source.media == 'twitter' || source.media == 'whatsapp' || source.media == 'crowdtangle' ">
+        <div class="form-group" ng-show="source.media == 'smsgh' || source.media == 'twitter' || source.media == 'whatsapp' || source.media == 'facebook' || source.media == 'instagram'">
           <label for="keywords" translate>Keywords</label>
           <textarea ng-model="source.keywords" class="form-control" id="keywords" name="keywords" placeholder="{{ 'Enter keywords' | translate }}" ng-required="source.media == 'twitter'">
           </textarea>
@@ -52,32 +52,12 @@
             <small class=" form-text text-muted" ng-show="form.keywords.$error.required">Keywords are required for Twitter sources.</small>
           </div>
         </div>
-
-        <div class="form-group" ng-show="source.media && source.media != 'twitter' && source.media != 'smsgh' && source.media != 'whatsapp' && source.media != 'crowdtangle' && source.media != 'comments'">
-          <label for="url" translate>URL</label>
-          <div class="input-group">
-            <div ng-switch="source.media">
-              <input type="url" ng-model="source.url" ng-switch-when="facebook" ng-pattern="/https?:\/\/www\.facebook\.com\//" class="form-control" id="url" name="url" ng-required="source.media != 'twitter' || source.media != 'comments'" ng-focus>
-              <input type="url" ng-model="source.url" ng-switch-default class="form-control" id="url" name="url" placeholder="{{ 'Enter url' | translate }}" ng-init="sourceUrl(source)" ng-required="!(source.media == 'twitter' || source.media == 'smsgh' || source.media == 'whatsapp' || source.media == 'crowdtangle' || source.media == 'comments')" ng-focus>
-            </div>
-            <div class="input-group-addon"><span ng-show="form.url.$focused && !form.url.$invalid" class="fa fa-check"></span></div>
-          </div>
-          <small class="form-text text-muted" ng-switch="source.media">
-            <span ng-switch-when="rss" translate="sources_modal.rss_url"></span>
-            <span ng-switch-when="facebook" translate="sources_modal.facebook_url"></span>
-            <span ng-switch-when="elmo" translate="sources_modal.elmo_url"></span>
-          </small>
-          <div class="has-error" ng-show="showErrors() && form.url.$invalid">
-            <small class="form-text text-muted" ng-show="form.url.$error.required" translate>URL is required.</small>
-            <small class="form-text text-muted" ng-show="form.url.$dirty && !form.url.$error.required" translate>URL is invalid.</small>
-          </div>
-        </div>
         <div class="form-group" ng-show="source.media && source.media != 'comments'">
           <label for="tags" translate>Tags</label>
           <input ng-model="source.tags" ng-list class="form-control" id="tags" name="tags" placeholder="{{ 'Enter tags' | translate }}">
           <small class="form-text text-muted" translate="sources_modal.tags"></small>
         </div>
-        <div class="form-group" ng-show="source.media && source.media == 'crowdtangle'">
+        <div class="form-group" ng-show="source.media && (source.media == 'facebook' || source.media == 'instagram')">
           <label for="dashboardAPIToken" translate>CrowdTangle Dashboard API Token</label>
           <input type="text" ng-model="source.dashboardAPIToken" class="form-control" id="dashboardAPIToken" name="dashboardAPIToken" placeholder="{{ 'Dashboard API Token' | translate }}">
         </div>
