+++ public/angular/templates/sources/show.html
@@ -43,7 +43,7 @@
               </tr>
               <tr>
                 <th scope="row" translate>Media</th>
-                <td class="detail-value">{{source.media | replaceCT | translate}}</td>
+                <td class="detail-value">{{source.media | translate}}</td>
               </tr>
               <tr>
                 <th scope="row" translate>Target</th>
