+++ public/angular/translations/locale-en.json
@@ -22,6 +22,7 @@
   "rss": "RSS",
   "elmo": "Elmo",
   "facebook": "Facebook",
+  "instagram": "Instagram",
   "smsgh": "SMS GH",
   "crowdtangle": "CrowdTangle",
   "comments": "Comments",
